requirements.txt
click==8.0.1
colorama==0.4.4
cmake==3.18.4.post1
dlib==19.21.1
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.1
gunicorn==20.1.0
importlib-metadata==4.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
numpy==1.20.0 
Pillow==8.2.0
scikit-learn==0.24.2
sklearn==0.0
srt==3.4.1
threadpoolctl==2.1.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.4.1
